support for selecting users who have the same name
Previously when there were multiple users who share the same
name a 404 was returned when trying to add them to a
project.

To fix this we record details about user chosen via the
ProjectTeamManager autocomplete popup so we can accurately
add them. The previous implementation only stored the text of
the name.

To allow users to tell a difference between users who share a
name the netid is added after the username in the user selection
autocomplete popup. This value is not retained in the textbox
for compatibility with existing code.